+++ snakebacon/mcmcbackends/__init__.py
@@ -1,0 +1,7 @@
+from .bacon import run_baconmcmc
+
+
+class Bacon:
+    def runmcmc(*args, **kwargs):
+        return run_baconmcmc(*args, **kwargs
+                             )